multi structure, separated all logic into the 2 services

+++ jelly-farm-service/src/main/java/aa/meijer/jelly/jellyFarmService/repository/JellyStockRepository.java
@@ -14,5 +14,5 @@ public interface JellyStockRepository extends JpaRepository<JellyEntity, UUID> {
     @Query("FROM JellyEntity j " +
             "WHERE j.dateTimeSold = NULL " +
             "AND j.cageNumber = :cageNumber")
-    List<JellyEntity> findByCageNumberUnsold(int cageNumber);
+    List<JellyEntity> findByCageNumberUnsold(long cageNumber);
 }
